Refactor setupEnvironmentHelper, VS Code install sequence, androidEmulatorHelper
