+++ test/smoke/src/helpers/iosSimulatorHelper.ts
@@ -1,3 +1,6 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for details.
+
 import { spawn } from "child_process";
 
 interface RunResult {
